refactor: refactor table id (#504)
* refactor tableId

* refactor raw table id

* meta service

* frontend

* fix

* clippy

+++ rust/meta/src/dashboard/mod.rs
@@ -121,7 +121,7 @@ mod handlers {
             .list_table_fragments()
             .map_err(err)?
             .iter()
-            .map(|f| (f.table_id().table_id(), f.actors()))
+            .map(|f| (f.table_id().table_id() as i32, f.actors()))
             .collect::<Vec<_>>();
 
         Ok(Json(table_fragments))
